refactor(breathing): major refactoring and PinkRain style integration
- Refactored breathing_screen.dart into modular components:
  * models.dart: State management and data models
  * exercises.dart: Exercise selection UI with Cupertino picker
  * player.dart: Active exercise display with animations
  * Reduced from ~900 lines to ~430 lines

- Updated UI to use PinkRain design system:
  * Replaced Material widgets with PinkRain Button components
  * Applied AppTokens for text styles and AppColors for palette
  * Replaced Material dropdown with Cupertino picker
  * Updated icons to HugeIcons (arrow up/down, pause, etc.)
  * Added custom SVG icons for exercises (box.svg, bow.svg, boat.svg)
  * Made buttons full-width and larger
  * Exercise descriptions are larger and bold

- UI/UX improvements:
  * White background during exercise selection
  * Animated gradient only appears when exercise is active
  * App bar hidden during exercise for immersive experience
  * Improved pause/resume button logic (always shows Pause when running)
